+++ src/App.jsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from 'styled-components'
 import './App.css'
 import { theme } from './utils/theme'
 import GlobalStyle from './styles/GlobalStyle'
-import { useEffect, useReducer, useState } from 'react'
+import { useEffect, useReducer, useRef, useState } from 'react'
 import { createContext } from 'react'
 import formatDate from './utils/formatDate'
 
@@ -12,6 +12,8 @@ import Editor from './components/Editor'
 import DailyContentHeader from './components/DailyContentHeader'
 import TodoViewer from './components/TodoViewer'
 import { loadTodos, saveTodos } from './utils/storage'
+import FlexContainerRow from './styles/FlexContainerRow'
+import WeeklyContentHeader from './components/WeeklyContentHeader'
 
 const todoReducer = (state, action) => {
   switch (action.type) {
@@ -95,19 +97,33 @@ function App() {
     })
   }
 
+  const dateInputRef = useRef(null)
+  const onClickDate = () => {
+    dateInputRef.current.showPicker()
+  }
+  const onChangeDate = (e) => {
+    setPivotDate(e.target.value)
+  }
+
   return (
     <>
       <TodoStateContext.Provider value={{ todos, pivotDate }}>
-        <TodoDispatchContext.Provider value={{ onCreate, onUpdateStatus, onDelete, setPivotDate }}>
+        <TodoDispatchContext.Provider
+          value={{ onCreate, onUpdateStatus, onDelete, setPivotDate, onClickDate }}>
           <ThemeProvider theme={theme}>
             <GlobalStyle />
             <Header />
-            <Container></Container>
-            <Container>
-              <DailyContentHeader />
-              <Editor />
-              <TodoViewer />
-            </Container>
+            <FlexContainerRow>
+              <Container>
+                <WeeklyContentHeader />
+              </Container>
+              <input ref={dateInputRef} onChange={onChangeDate} type="date" hidden />
+              <Container>
+                <DailyContentHeader />
+                <Editor />
+                <TodoViewer />
+              </Container>
+            </FlexContainerRow>
           </ThemeProvider>
         </TodoDispatchContext.Provider>
       </TodoStateContext.Provider>
